// src/pbERP/Web/src/app/shared/models/mainMenu.ts
// export interface MainMenu {
//   companyLinkModuleId: number
//   companyId: number
//   moduleId: number
//   companyName: string
//   moduleName: string
// }

export interface MainMenu {
  screenId: number
  screenName: string
  controllerName: string
  actionName: string
  moduleId: number
  parentId: number
}